Don't commit webpack-generated bundles.
First this removes the already commited JS/CSS files in
src/main/webapp from the repository.

Then since everything in src/main/webapp is generated by webpack
at the moment we just gitignore everything in that directory. If any
static assets are added in the future we can always be more specific.