Show both future and past events in calendar view.
Future events are scheduled job exceutions and their duration
is estimated based on past jobs executions.

Past events are prior runs from a jobs build history. Their
duration is the actual duration of the past job execution.

Colors for now are:
- blue for future builds
- red for failed past builds
- green for successful past builds
- yellow for unstable past builds
- gray for other past builds

These are non-standard colors for Jenkins (green instead of
blue for success). Might be better to find a different color for
future events and stick to the standard blue or maybe just
make colors configurable.