Fix calculation of next start date
CronTab.ceil() modifies its argument. That means if there
is more than one trigger spec the calculated starting date
will be wrong. We need to clone the current date so we
can reuse it in the next iteration.

This commit also refactors out the current date into a method
argument to make testing easier.